+++ src/SadConsole.Shared/Input/MouseConsoleState.cs
@@ -39,27 +39,31 @@ namespace SadConsole.Input
             {
                 if (console.UsePixelPositioning)
                 {
-                    RelativePixelPosition = mouseData.ScreenPosition - console.RelativePosition;
+                    RelativePixelPosition = mouseData.ScreenPosition - console.CalculatedPosition;
                     WorldPosition = mouseData.ScreenPosition;
                     ConsolePosition = RelativePixelPosition.PixelLocationToConsole(console.TextSurface.Font);
-                    IsOnConsole = new Rectangle(0, 0, console.TextSurface.Width, console.TextSurface.Height).Contains(ConsolePosition);
+
+                    var tempCellPosition = ConsolePosition + Console.TextSurface.RenderArea.Location;
+                    IsOnConsole = Console.TextSurface.RenderArea.Contains(tempCellPosition);
 
                     if (IsOnConsole)
                     {
-                        CellPosition = ConsolePosition + Console.TextSurface.RenderArea.Location;
+                        CellPosition = tempCellPosition;
                         Cell = console.TextSurface[CellPosition.X, CellPosition.Y];
                     }
                 }
                 else
                 {
-                    RelativePixelPosition = mouseData.ScreenPosition - console.RelativePosition.ConsoleLocationToPixel(console.TextSurface.Font);
+                    RelativePixelPosition = mouseData.ScreenPosition - console.CalculatedPosition.ConsoleLocationToPixel(console.TextSurface.Font);
                     WorldPosition = mouseData.ScreenPosition.PixelLocationToConsole(console.TextSurface.Font);
-                    ConsolePosition = WorldPosition - console.RelativePosition;
-                    IsOnConsole = new Rectangle(0, 0, console.TextSurface.Width, console.TextSurface.Height).Contains(ConsolePosition);
+                    ConsolePosition = WorldPosition - console.CalculatedPosition;
+
+                    var tempCellPosition = ConsolePosition + Console.TextSurface.RenderArea.Location;
+                    IsOnConsole = Console.TextSurface.RenderArea.Contains(tempCellPosition);
 
                     if (IsOnConsole)
                     {
-                        CellPosition = ConsolePosition + Console.TextSurface.RenderArea.Location;
+                        CellPosition = tempCellPosition;
                         Cell = console.TextSurface[CellPosition.X, CellPosition.Y];
                     }
                 }